+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [0.3.1][] (2022-07-31)
+
+## Fixed
+
+- Frozen strings returned from providers caused an error.
+
 # [0.3.0][] (2022-07-31)
 
 ## Changed
@@ -21,6 +27,7 @@
 
 Initial release.
 
+[0.3.1]: https://github.com/AaronLasseigne/unifig/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/AaronLasseigne/unifig/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/AaronLasseigne/unifig/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/AaronLasseigne/unifig/compare/v0.0.0...v0.1.0